+++ libraries/MAX471_RT/MAX471.h
@@ -2,7 +2,7 @@
 //
 //    FILE: MAX471.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 //    DATE: 2024-01-30
 // PURPOSE: Arduino library for MAX471 current sensor.
 //     URL: https://github.com/RobTillaart/MAX471_RT
@@ -11,7 +11,7 @@
 
 #include "Arduino.h"
 
-#define MAX471_LIB_VERSION        (F("0.1.1"))
+#define MAX471_LIB_VERSION        (F("0.1.2"))
 
 
 class MAX471
@@ -24,14 +24,25 @@ public:
 
   //  ADC parameters, defaults 10 bit 5V (UNO)
   //  assumes both ADC are equal.
-  void      begin(float maxVoltage = 5, uint16_t maxSteps = 1023);
+  void  begin(float maxVoltage = 5.0, uint16_t maxSteps = 1023);
 
-  float     readCurrent(uint8_t times = 1);
-  float     readCurrentMilliAmpere(uint8_t times = 1);
-  float     readVoltage(uint8_t times = 1);
-  float     readVoltageMilliVolts(uint8_t times = 1);
-  float     calcPower();
-  float     calcPowerMilliWatt();
+  //  read the sensor
+  float readCurrent(uint8_t times = 1);
+  float readCurrentMilliAmpere(uint8_t times = 1);
+  float readVoltage(uint8_t times = 1);
+  float readVoltageMilliVolts(uint8_t times = 1);
+  //  power = last current x last voltage
+  float calcPower();
+  float calcPowerMilliWatt();
+
+  //  read cached values.
+  float getLastCurrent();
+  float getLastVoltage();
+
+  //  experimental
+  //  overrules the need for readVoltage()
+  //  by setting a fixed voltage.
+  void      setFixedVoltage(float volts);
 
 
 protected:
@@ -40,7 +51,19 @@ protected:
   uint8_t  _signPin;
   float    _current;
   float    _voltage;
-  float    _units;
+  float    _units;  //  volts per step
+};
+
+
+//////////////////////////////////////////////////////////////////
+//
+//  DERIVED CLASSES
+//
+class MAX472 : public MAX471
+{
+public:
+  MAX472(uint8_t currentPin, uint8_t voltagePin);
+  MAX472(uint8_t currentPin, uint8_t voltagePin, uint8_t signPin);
 };