+++ libraries/MAX471_RT/CHANGELOG.md
@@ -6,5 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2025-11-21
+- improve documentation
+- add **setFixedVoltage(float v)**
+- add **float getLastCurrent()**
+- add **float getLastVoltage()**
+- add MAX472 derived class (wraps MAX471)
+- update GitHub actions
+- minor edits
+
+## [0.1.1] - 2024-01-30
+- fix platformIO link
+
 ## [0.1.0] - 2024-01-30
 - initial version